Test: fix power-of-two data type in uint64_t division tests

+++ test/Quotient.cc
@@ -149,7 +149,7 @@ TEST(uint64_t, divide_by_n_minus_1) {
 
 TEST(uint64_t, divide_by_power_of_2) {
 	const uint64_t step = UINT64_C(93400375993241);
-	for (uint64_t p = 0; p < 64; p += 1) {
+	for (uint32_t p = 0; p < 64; p += 1) {
 		const fxdiv_divisor_uint64_t divisor =
 			fxdiv_init_uint64_t(UINT64_C(1) << p);
 		for (uint64_t n = 0; n <= UINT64_MAX - step + 1; n += step) {
@@ -160,7 +160,7 @@ TEST(uint64_t, divide_by_power_of_2) {
 
 TEST(uint64_t, divide_by_power_of_2_minus_1) {
 	const uint64_t step = UINT64_C(93400375993241);
-	for (uint64_t p = 0; p < 64; p += 1) {
+	for (uint32_t p = 0; p < 64; p += 1) {
 		const uint64_t d = (UINT64_C(2) << p) - UINT64_C(1);
 		const fxdiv_divisor_uint64_t divisor = fxdiv_init_uint64_t(d);
 		for (uint64_t n = 0; n <= UINT64_MAX - step + 1; n += step) {
